Change 'trait_category' to String from Double.
Fix schema error in V2D index. When the pipeline is run using Parquet
formats, the output of V2D is saved with `trait_category` as a String.
When this is read back in in subsequent steps, Spark tries to convert
the BYTE_ARRAY into a Double as specified by the `schema` and the
pipeline breaks. This works in JSON because it is so permissible.